refactor: simplify booting logic
The abstraction of BackgroudService don't seem to be as important as before now that there are
only two of said services. Also they are quire rigid about how they share data, which is a problem
for the new architecture that we are going through where more of the infra creation will be on the
local server and then shared.

This hopefully simplifies it. The services are simple function calls that we can expand on its
params and that their params are not tied into the same constraints.

Also as part of this, removed the dynamic loading for booting. this also adds to the simplification
of it, but is not a blocker if we want to add the loading again instead of a "progress logging".
