+++ linkup-cli/src/commands/local.rs
@@ -2,8 +2,7 @@ use anyhow::anyhow;
 use colored::Colorize;
 
 use crate::{
-    Result,
-    services::{self, BackgroundService},
+    Result, services,
     state::{ServiceTarget, State, upload_state},
 };
 
@@ -35,7 +34,7 @@ pub async fn local(args: &Args) -> Result<()> {
         return Ok(());
     }
 
-    if services::LocalServer::find_pid().is_none() {
+    if services::local_server::find_pid().is_none() {
         println!(
             "{}",
             "Seems like your local Linkup server is not running. Please run 'linkup start' first."
